Computing Circularity with Perimeter algorithm

+++ src/main/java/segmentBlobs/GlobalThresholding.java
@@ -50,7 +50,7 @@ package segmentBlobs;
 			}
 			
 
-			return ThresholdNew;
+			return (float)ThresholdNew;
 
 		}
